MC 1.14 compatibility and miscellaneous improvements and fixes This change adds 1.14 Spigot API and server compatibility, by not using methods that were deprecated in 1.13 and subsequently removed (it only affected getTitle() method calls on Inventory in ColorGUIListener).
It also deletes bStats vendor code, replacing it by a simple POM dependency that
should be easier to mantain in the future and reduce build times.
The new bStats artifact depends on the 1.13.2 Spigot API version, but I didn't
see that it used any method not available on previous API versions, so it shouldn't
break compatibility with older servers. I didn't test this, however, so any report about
this change breaking compatibility is welcome.

I've also removed an apparent debug broadcastMessage method call that somehow made it to
this repo. I don't think that all players are interested in knowing the modifiers reapplied after
one changes the chat color by using the GUI.

Finally, I removed some unused includes, and bumped the plugin version to 1.7.8.