Amended the group colours system to have a predictable selection system when a player has more than one group permission.
Now, the order that the groups are placed in the config is the order that they will be applied in-game. i.e., the top-level group will take priority over the other groups. This means that group hierarchies will now work correctly.
Amended several IDE warnings, functionally nothing should have changed.